sepet

+++ frontend/src/components/areUSure.js
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { Button, Popconfirm } from 'antd';
-import { DeleteOutlined } from '@ant-design/icons';
-const ConfirmPopUp = ({handleDelete , id}) => {
+import { CheckCircleOutlined, DeleteOutlined } from '@ant-design/icons';
+const ConfirmPopUp = ({ handleDelete, handleConfirm, isConfirm, id, record }) => {
   const [open, setOpen] = useState(false);
   const [confirmLoading, setConfirmLoading] = useState(false);
   const showPopconfirm = () => {
@@ -12,7 +12,13 @@ const ConfirmPopUp = ({handleDelete , id}) => {
     setTimeout(() => {
       setOpen(false);
       setConfirmLoading(false);
-      handleDelete(id);
+      if (isConfirm) {
+        handleConfirm(record);
+      }
+      else {
+        handleDelete(id);
+      }
+
       console.log(id);
     }, 2000);
   };
@@ -21,25 +27,53 @@ const ConfirmPopUp = ({handleDelete , id}) => {
     setOpen(false);
   };
   return (
-    <Popconfirm
-      title="Kaydı Sil"
-      description="Bu kaydı silmek istediğinize emin misiniz?"
-      open={open}
-      c
-      onConfirm={handleOk}
-      okButtonProps={{
-        loading: confirmLoading,
-        danger:true,
-      }}
-      onCancel={handleCancel}
-    >
-      <Button
-        type="link"
-        icon={<DeleteOutlined />}
-        onClick={showPopconfirm}
-      >
-      </Button>
-    </Popconfirm>
+    <>
+      {isConfirm ? (
+        <Popconfirm
+          title="Kaydı Onayla ve Mail Gönder"
+          description="Bu kaydı onaylamak istediğinize emin misiniz?"
+          open={open}
+          okText="Onayla"
+          cancelText="İptal"
+          onConfirm={handleOk}
+
+          okButtonProps={{
+            loading: confirmLoading,
+            danger: false,
+            style: { backgroundColor: 'green' }
+          }}
+
+          onCancel={handleCancel}
+        >
+          <Button
+            type="link"
+            icon={<CheckCircleOutlined />}
+            onClick={showPopconfirm}
+          />
+        </Popconfirm>
+      ) : (
+        <Popconfirm
+          title="Kaydı Sil"
+          description="Bu kaydı silmek istediğinize emin misiniz?"
+          open={open}
+          okText="Sil"
+          cancelText="İptal"
+          onConfirm={handleOk}
+          okButtonProps={{
+            loading: confirmLoading,
+            danger: true,
+          }}
+          onCancel={handleCancel}
+        >
+          <Button
+            type="link"
+            icon={<DeleteOutlined />}
+            onClick={showPopconfirm}
+          />
+        </Popconfirm>
+      )}
+    </>
+
   );
 };
 export default ConfirmPopUp;
